[CoordinatedGraphics] Move CoordinatedGraphicsBackingStoreTile to its own file https://bugs.webkit.org/show_bug.cgi?id=281488
Reviewed by Miguel Gomez.

And simplify it. We don't really need to inherit from
TextureMapperTile, since we are using it just as a wrapper of a
BitmapTexture.

* Source/WebCore/platform/TextureMapper.cmake:
* Source/WebCore/platform/graphics/texmap/coordinated/CoordinatedBackingStore.cpp:
(WebCore::CoordinatedBackingStore::processPendingUpdates):
(WebCore::CoordinatedBackingStore::resize):
(WebCore::CoordinatedBackingStore::paintToTextureMapper):
(WebCore::CoordinatedBackingStore::drawBorder):
(WebCore::CoordinatedBackingStore::drawRepaintCounter):
(WebCore::CoordinatedBackingStoreTile::addUpdate): Deleted.
(WebCore::CoordinatedBackingStoreTile::swapBuffers): Deleted.
(WebCore::CoordinatedBackingStore::paintTilesToTextureMapper): Deleted.
(WebCore::CoordinatedBackingStore::adjustedTransformForRect): Deleted.
(WebCore::CoordinatedBackingStore::swapBuffers): Deleted.
* Source/WebCore/platform/graphics/texmap/coordinated/CoordinatedBackingStore.h:
* Source/WebCore/platform/graphics/texmap/coordinated/CoordinatedBackingStoreTile.cpp: Added.
(WebCore::CoordinatedBackingStoreTile::CoordinatedBackingStoreTile):
(WebCore::CoordinatedBackingStoreTile::addUpdate):
(WebCore::CoordinatedBackingStoreTile::processPendingUpdates):
* Source/WebCore/platform/graphics/texmap/coordinated/CoordinatedBackingStoreTile.h: Added.
(WebCore::CoordinatedBackingStoreTile::texture):
(WebCore::CoordinatedBackingStoreTile::scale const):
(WebCore::CoordinatedBackingStoreTile::rect):
(WebCore::CoordinatedBackingStoreTile::canBePainted const):
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/CoordinatedGraphicsScene.cpp:
(WebKit::CoordinatedGraphicsScene::updateSceneState):

Canonical link: https://commits.webkit.org/285252@main
